+++ backend/app/api/v2/home/quotes.py
